+++ src/api/books/book.interface.ts
