+++ src/api/books/book.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 // from app
-import { BookService } from '@/book/book.service';
-import { BookController } from '@/book/book.controller';
-import { Book } from '@/book/book.entity';
+import { BookService } from '@/api/books/book.service';
+import { BookController } from '@/api/books/book.controller';
+import { Book } from '@/api/books/book.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Book])],
